Make manifest pages draggable
