+++ app/lib/widgets/components/app/entries_graph.dart
@@ -1,23 +1,15 @@
 import "package:flutter/material.dart";
-import "package:flutter_hooks/flutter_hooks.dart";
-import "package:graphview/GraphView.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 import "package:typewriter/models/entry.dart";
 import "package:typewriter/models/entry_blueprint.dart";
 import "package:typewriter/models/page.dart";
 import "package:typewriter/pages/page_editor.dart";
-import "package:typewriter/widgets/components/app/empty_screen.dart";
-import "package:typewriter/widgets/components/app/entry_node.dart";
+import "package:typewriter/widgets/components/app/draggable_graph.dart";
 import "package:typewriter/widgets/components/app/entry_search.dart";
 import "package:typewriter/widgets/components/app/search_bar.dart";
-import "package:typewriter/utils/passing_reference.dart";
-import 'dart:async';
-import 'dart:ui';
 
 part "entries_graph.g.dart";
-
-const double kVirtualCanvasSize = 200000;
 
 @riverpod
 List<Entry> graphableEntries(Ref ref) {
@@ -85,35 +77,6 @@ Set<String>? entryTriggers(Ref ref, String entryId) {
       .toSet();
 }
 
-@riverpod
-Graph graph(Ref ref) {
-  final entries = ref.watch(graphableEntriesProvider);
-  final graph = Graph();
-
-  for (final entry in entries) {
-    final node = Node.Id(entry.id);
-    graph.addNode(node);
-  }
-
-  for (final entry in entries) {
-    final triggeredEntryIds = ref.watch(entryTriggersProvider(entry.id));
-    if (triggeredEntryIds == null) continue;
-
-    final color = ref.watch(entryBlueprintProvider(entry.blueprintId))?.color ??
-        Colors.grey;
-
-    for (final triggeredEntryId in triggeredEntryIds) {
-      if (triggeredEntryId == entry.id) continue;
-      graph.addEdge(
-        Node.Id(entry.id),
-        Node.Id(triggeredEntryId),
-        paint: Paint()..color = color,
-      );
-    }
-  }
-
-  return graph;
-}
 
 @riverpod
 Offset nodePosition(Ref ref, String nodeId) {
@@ -126,283 +89,34 @@ Offset nodePosition(Ref ref, String nodeId) {
   return page.nodePositions[nodeId] ?? Offset.zero;
 }
 
-class EntriesGraph extends ConsumerStatefulWidget {
+@riverpod
+Map<String, Set<String>> triggerEdges(Ref ref) {
+  final Map<String, Set<String>> edges = {};
+  for (final entry in ref.watch(graphableEntriesProvider)) {
+    final triggeredIds = ref.watch(entryTriggersProvider(entry.id)) ?? {};
+    edges[entry.id] = triggeredIds;
+  }
+  return edges;
+}
+
+class EntriesGraph extends HookConsumerWidget {
   const EntriesGraph({super.key});
 
   @override
-  ConsumerState<EntriesGraph> createState() => _EntriesGraphState();
-}
-
-class _EntriesGraphState extends ConsumerState<EntriesGraph> with SingleTickerProviderStateMixin {
-  final TransformationController _controller = TransformationController();
-  Rect _visibleRect = Rect.zero;
-  List<String> _visibleNodes = [];
-  Timer? _throttleTimer;
-  late final AnimationController _animController;
-
-  @override
-  void initState() {
-    super.initState();
-
-    // Setup anim controller for animated edges
-    _animController = AnimationController(
-      vsync: this,
-      duration: const Duration(seconds: 1),
-    )..repeat(); // continuously loops
-
-    // Listen for viewport changes (pan/zoom)
-    _controller.addListener(_onViewportChanged);
-
-    // Initialize viewport culling immediately
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _onViewportChanged();
-      _initializeNodePositions();
-    });
-  }
-
-  @override
-  void dispose() {
-    debugPrint("Disposing GraphEntries");
-
-    _controller.removeListener(_onViewportChanged);
-    _controller.dispose();
-    _throttleTimer?.cancel();
-    _animController.dispose();
-    super.dispose();
-  }
-
-  void _initializeNodePositions() {
-    final page = ref.read(currentPageProvider);
-    if (page == null) return;
-
-    final entryIds = ref.read(graphableEntryIdsProvider);
-    for (int i = 0; i < entryIds.length; i++) {
-      final id = entryIds[i];
-      final pos = page.nodePositions[id];
-      if (pos == null) {
-        final initialOffset = Offset(100.0 + 200.0 * i, 100.0);
-        page.updateNodePosition(ref.passing, id, initialOffset);
-      }
-    }
-  }
-
-  void _onViewportChanged() {
-    // Throttle viewport changes to avoid flooding rebuilds
-    if (_throttleTimer?.isActive ?? false) return;
-    _throttleTimer = Timer(const Duration(milliseconds: 50), () {
-      final matrix = _controller.value;
-      final screenSize = MediaQuery
-          .of(context)
-          .size;
-
-      // Screen rect in screen space
-      final screenRect = Rect.fromLTWH(
-          0, 0, screenSize.width, screenSize.height);
-
-      // Transform into world/graph space
-      final inverseMatrix = Matrix4.inverted(matrix);
-      final worldTopLeft = MatrixUtils.transformPoint(
-          inverseMatrix, screenRect.topLeft);
-      final worldBottomRight = MatrixUtils.transformPoint(
-          inverseMatrix, screenRect.bottomRight);
-
-      // This is the new rect we will use for viewport culling
-      final newVisibleRect = Rect.fromPoints(worldTopLeft, worldBottomRight);
-
-      // Compute the nodes we can see relative to the viewport
-      final entryIds = ref.read(graphableEntryIdsProvider);
-      final newVisibleNodes = entryIds.where((id) {
-        final pos = ref.read(nodePositionProvider(id));
-        final nodeRect = Rect.fromLTWH(pos.dx, pos.dy, 120,
-            60); // TODO: Don't make constant height (120, 60), get actual node dimensions
-        return newVisibleRect.overlaps(nodeRect);
-      }).toList();
-
-      setState(() {
-        _visibleRect = newVisibleRect;
-        _visibleNodes = newVisibleNodes;
-      });
-
-      debugPrint("Visible rect updated: $_visibleRect");
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     final entryIds = ref.watch(graphableEntryIdsProvider);
+    final edges = ref.watch(triggerEdgesProvider);
 
-    if (entryIds.isEmpty) {
-      return EmptyScreen(
-        title: "There are no graphable entries on this page.",
-        buttonText: "Add Entry",
-        onButtonPressed: () => ref.read(searchProvider.notifier).asBuilder()
-          ..fetchNewEntry()
-          ..nonGenericAddEntry()
-          ..tag("trigger")
-          ..open(),
-      );
-    }
-
-    final Map<String, Set<String>> edges = {};
-    for (final entry in ref.watch(graphableEntriesProvider)) {
-      final triggeredIds = ref.watch(entryTriggersProvider(entry.id)) ?? {};
-      edges[entry.id] = triggeredIds;
-    }
-
-    return InteractiveViewer(
-      transformationController: _controller,
-      constrained: false,
-      boundaryMargin: const EdgeInsets.all(1000),
-      minScale: 0.1,
-      maxScale: 2.5,
-      child: SizedBox(
-        width: kVirtualCanvasSize,
-        height: kVirtualCanvasSize,
-        child: Stack(
-          children: [
-            // Paint edges only if their endpoints are visible
-            AnimatedBuilder(
-              animation: _animController,
-              builder: (context, child) {
-                return CustomPaint(
-                  painter: EdgePainter(
-                  positions: entryIds
-                      .map((id) => MapEntry(id, ref.watch(nodePositionProvider(id))))
-                      .toMap(),
-                  edges: edges,
-                  visibleNodes: _visibleNodes,
-                  dashOffset: _animController.value,
-                  ),
-                );
-              },
-            ),
-            // Render only visible nodes
-            ..._visibleNodes.map((id) {
-              return Consumer( // Wrap in consumer to isolate the provider, meaning when we drag a node, the entire graph does not need to update
-                builder: (context, ref, _) {
-                  final position = ref.watch(nodePositionProvider(id));
-
-                  return Positioned(
-                    left: position.dx,
-                    top: position.dy,
-                    child: GestureDetector(
-                      onPanUpdate: (details) {
-                        final newPos = Offset(
-                          (position.dx + details.delta.dx).clamp(0.0, double.infinity),
-                          (position.dy + details.delta.dy).clamp(0.0, double.infinity),
-                        );
-                        // Immediately update local position for smooth dragging
-                        final page = ref.read(currentPageProvider);
-                        if (page != null) {
-                          page.syncUpdateNodePosition(ref.passing, id, newPos);
-                        }
-                      },
-                      onPanEnd: (details) {
-                        // Since the client who is panning gets predicted updates, we can update all the other clients on pan end
-                        final position = ref.read(nodePositionProvider(id));
-
-                        final page = ref.read(currentPageProvider);
-                        if (page != null) {
-                          page.updateNodePosition(ref.passing, id, position);
-                        }
-                      },
-                      child: EntryNode(entryId: id, key: ValueKey(id)),
-                    ),
-                  );
-                },
-              );
-            }),
-          ],
-        ),
-      ),
+    return DraggableGraph(
+      entryIds: entryIds,
+      edges: edges,
+      emptyTitle: "There are no graphable entries on this page.",
+      emptyButtonText: "Add Entry",
+      onEmptyButtonPressed: () => ref.read(searchProvider.notifier).asBuilder()
+        ..fetchNewEntry()
+        ..nonGenericAddEntry()
+        ..tag("trigger")
+        ..open(),
     );
-  }
-}
-
-class EdgePainter extends CustomPainter {
-  final Map<String, Offset> positions;
-  final Map<String, Set<String>> edges;
-  final List<String> visibleNodes;
-  final double dashOffset; // 0..1
-  final Color color;
-
-  EdgePainter({
-    required this.positions,
-    required this.edges,
-    required this.visibleNodes,
-    required this.dashOffset,
-    this.color = Colors.green,
-  });
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = color
-      ..strokeWidth = 2
-      ..style = PaintingStyle.stroke;
-
-    final edgeOffset = Offset(100, 25);
-    final arrowSize = 6.0; // length of arrow lines
-    final arrowWidth = 7.0; // angle at tip
-    final arrowSpacing = 20.0; // distance between arrows along edge
-    final arrowSpeed = 2.0;
-
-    final points = <Offset>[];
-
-    edges.forEach((fromId, toIds) {
-      final from = (positions[fromId] ?? Offset.zero)  + edgeOffset;
-      if (from == null) return;
-
-      final visibleFrom = visibleNodes.contains(fromId);
-
-      for (final toId in toIds) {
-        if (!visibleNodes.contains(toId) && !visibleFrom) continue;
-
-        final to = (positions[toId] ?? Offset.zero) + edgeOffset;
-        if (to == null) continue;
-
-        final direction = to - from;
-        final length = direction.distance;
-        if (length == 0) continue;
-
-        final unit = direction / length;
-        final perp = Offset(-unit.dy, unit.dx) * (arrowWidth / 2);
-
-        final totalArrows = (length / arrowSpacing).ceil();
-
-        for (int i = 0; i < totalArrows; i++) {
-          // Animated offset along the edge
-          final t = ((i * arrowSpacing) + dashOffset * arrowSpacing * arrowSpeed) % length;
-          final center = from + unit * t;
-
-          final tip = center + unit * arrowSize;
-
-          // Add two line segments for the arrowhead
-          points.add(tip);
-          points.add(center + perp);
-
-          points.add(tip);
-          points.add(center - perp);
-        }
-      }
-    });
-
-    // Batch draw calls
-    canvas.drawPoints(PointMode.lines, points, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant EdgePainter oldDelegate) {
-    return positions != oldDelegate.positions ||
-        edges != oldDelegate.edges ||
-        visibleNodes != oldDelegate.visibleNodes ||
-        (dashOffset - oldDelegate.dashOffset).abs() > 0.01;
-  }
-}
-
-// ------------------- Extension to convert Iterable<MapEntry> to Map -------------------
-extension ToMap<K, V> on Iterable<MapEntry<K, V>> {
-  Map<K, V> toMap() {
-    return Map.fromEntries(this);
   }
 }
