+++ app/lib/widgets/components/app/manifest_view.dart
@@ -1,14 +1,11 @@
 import "package:flutter/material.dart";
-import "package:flutter_hooks/flutter_hooks.dart";
-import "package:graphview/GraphView.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 import "package:typewriter/models/entry.dart";
 import "package:typewriter/models/entry_blueprint.dart";
 import "package:typewriter/models/page.dart";
 import "package:typewriter/pages/page_editor.dart";
-import "package:typewriter/widgets/components/app/empty_screen.dart";
-import "package:typewriter/widgets/components/app/entry_node.dart";
+import "package:typewriter/widgets/components/app/draggable_graph.dart";
 import "package:typewriter/widgets/components/app/entry_search.dart";
 import "package:typewriter/widgets/components/app/search_bar.dart";
 
@@ -59,38 +56,43 @@ Set<String>? entryReferences(Ref ref, String entryId) {
       .toSet();
 }
 
-@riverpod
-Graph manifestGraph(Ref ref) {
-  final entries = ref.watch(manifestEntriesProvider);
-  final graph = Graph();
 
-  for (final entry in entries) {
-    final node = Node.Id(entry.id);
-    graph.addNode(node);
+@riverpod
+Offset manifestNodePosition(Ref ref, String nodeId) {
+  final page = ref.watch(currentPageProvider);
+  if (page == null) {
+    debugPrint("Page was null when accessing position for node. This is a bug.");
+    return Offset.zero;
   }
+
+  return page.nodePositions[nodeId] ?? Offset.zero;
+}
+
+@riverpod
+Map<String, Set<String>> manifestEdges(Ref ref) {
+  final entries = ref.watch(manifestEntriesProvider);
+  final Map<String, Set<String>> edges = {};
 
   for (final entry in entries) {
     final referenceEntryIds = ref.watch(entryReferencesProvider(entry.id));
-    if (referenceEntryIds == null) continue;
-
-    final color = ref.watch(entryBlueprintProvider(entry.blueprintId))?.color ??
-        Colors.grey;
-
-    for (final referenceEntryId in referenceEntryIds) {
-      final referenceTags = ref.watch(entryTagsProvider(referenceEntryId));
-      // We only want to show references to manifest entries.
-      if (!referenceTags.contains("manifest")) continue;
-      if (referenceEntryId == entry.id) continue;
-
-      graph.addEdge(
-        Node.Id(entry.id),
-        Node.Id(referenceEntryId),
-        paint: Paint()..color = color,
-      );
+    if (referenceEntryIds == null) {
+      edges[entry.id] = {};
+      continue;
     }
+
+    // Only include references to manifest entries that are on the current page
+    final entryIds = ref.watch(manifestEntryIdsProvider);
+    final manifestReferences = referenceEntryIds.where((referenceEntryId) {
+      final referenceTags = ref.watch(entryTagsProvider(referenceEntryId));
+      return referenceTags.contains("manifest") &&
+             referenceEntryId != entry.id &&
+             entryIds.contains(referenceEntryId);
+    }).toSet();
+
+    edges[entry.id] = manifestReferences;
   }
 
-  return graph;
+  return edges;
 }
 
 class ManifestView extends HookConsumerWidget {
@@ -99,66 +101,18 @@ class ManifestView extends HookConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final entryIds = ref.watch(manifestEntryIdsProvider);
-    final graph = ref.watch(manifestGraphProvider);
+    final edges = ref.watch(manifestEdgesProvider);
 
-    final builder = useMemoized(
-      () => SugiyamaConfiguration()
-        ..nodeSeparation = 40
-        ..levelSeparation = 40
-        ..orientation = SugiyamaConfiguration.ORIENTATION_TOP_BOTTOM,
-    );
-
-    if (entryIds.isEmpty) {
-      return EmptyScreen(
-        title: "There are no manifest entries on this page.",
-        buttonText: "Add Entry",
-        onButtonPressed: () => ref.read(searchProvider.notifier).asBuilder()
-          ..fetchNewEntry()
-          ..nonGenericAddEntry()
-          ..tag("manifest")
-          ..open(),
-      );
-    }
-
-    return InteractiveViewer(
-      constrained: false,
-      boundaryMargin: EdgeInsets.symmetric(
-        horizontal: MediaQuery.of(context).size.width,
-        vertical: MediaQuery.of(context).size.height,
-      ),
-      minScale: 0.0001,
-      maxScale: 2.6,
-      child: GraphView(
-        graph: graph,
-        algorithm: SugiyamaAlgorithm(builder),
-        paint: Paint()
-          ..color = Colors.green
-          ..strokeWidth = 1
-          ..style = PaintingStyle.stroke,
-        builder: (node) {
-          final id = node.key?.value as String?;
-          if (id == null) return const NonExistentEntry();
-
-          final entryOnPage = entryIds.contains(id);
-          if (!entryOnPage) {
-            final globalEntryWithPage =
-                ref.watch(globalEntryWithPageProvider(id));
-            if (globalEntryWithPage == null) {
-              return const NonExistentEntry();
-            }
-
-            return ExternalEntryNode(
-              pageId: globalEntryWithPage.key,
-              entry: globalEntryWithPage.value,
-            );
-          }
-
-          return EntryNode(
-            entryId: id,
-            key: ValueKey(id),
-          );
-        },
-      ),
+    return DraggableGraph(
+      entryIds: entryIds,
+      edges: edges,
+      emptyTitle: "There are no manifest entries on this page.",
+      emptyButtonText: "Add Entry",
+      onEmptyButtonPressed: () => ref.read(searchProvider.notifier).asBuilder()
+        ..fetchNewEntry()
+        ..nonGenericAddEntry()
+        ..tag("manifest")
+        ..open(),
     );
   }
 }
